Advance Rag Step Back Prompting System/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-qdrant
langchain-core
qdrant-client
python-dotenv
pypdf